fix - 리뷰어님 피드백 반영 2
- 객체 출력 시 domain.toString() 대신에 view 에서 로직을 구현하도록 이동 (LottoTicket)
- swich-case 문 대신 stream을 이용하도록 변경 (LottoMatchCaseEnum)
- 일반 메소드가 생성자 다음에 오도록 위치 변경 (LottoNumber)
- LottoTicketsPrice, LottoWinningPrice 두 클래스를 Money 클래스로 추출

+++ src/main/java/step3/view/InputView.java
@@ -25,8 +25,8 @@ public class InputView {
         .collect(Collectors.toList());
   }
 
-  public static int getPurchasePriceInput() {
+  public static Long getPurchasePriceInput() {
     ResultView.printPurchasePriceRequest();
-    return Integer.parseInt(sc.nextLine());
+    return Long.parseLong(sc.nextLine());
   }
 }
